Makefile: adjust to new Homebrew Cask behaviours.
Homebrew Cask now moves rather than copies .app bundles so we no longer
need to look for the location of the app. As a bonus this will also
handle the case where it was installed outside Homebrew Cask.

Additionally, stick a `brew update` in here just to make sure folks are
on the latest Homebrew Cask version with this behaviour.